chore: move addPath function
