+++ src/utils/setup/setupPipPack.ts
@@ -3,7 +3,7 @@ import { getExecOutput } from "@actions/exec"
 import execa from "execa"
 import which from "which"
 import { info } from "@actions/core"
-import { addPath } from "../path/addPath"
+import { addPath } from "../env/addEnv"
 import { setupPython } from "../../python/python"
 import { isBinUptoDate } from "./version"
 import { join } from "path"
